e-commerce-system-backend/requirements.txt
django==4.2
djangorestframework
django-cors-headers
mysqlclient
cloudinary
python-decouple